convert keyhashes to bech32 and implement rollback for rewardbyaddress

+++ src/Conclave.Sink/Reducers/DelegatorByEpochReducer.cs
@@ -1,6 +1,11 @@
+using CardanoSharp.Wallet.Enums;
+using CardanoSharp.Wallet.Extensions;
+using CardanoSharp.Wallet.Utilities;
 using Conclave.Sink.Data;
 using Conclave.Sink.Models;
+using Conclave.Sink.Services;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Options;
 
 namespace Conclave.Sink.Reducers;
 
@@ -9,10 +14,17 @@ public class DelegatorByEpochReducer : OuraReducerBase
 {
     private readonly ILogger<DelegatorByEpochReducer> _logger;
     private IDbContextFactory<ConclaveSinkDbContext> _dbContextFactory;
-    public DelegatorByEpochReducer(ILogger<DelegatorByEpochReducer> logger, IDbContextFactory<ConclaveSinkDbContext> dbContextFactory)
+    private readonly CardanoService _cardanoService;
+    private readonly ConclaveSinkSettings _settings;
+    public DelegatorByEpochReducer(ILogger<DelegatorByEpochReducer> logger,
+        IDbContextFactory<ConclaveSinkDbContext> dbContextFactory,
+        IOptions<ConclaveSinkSettings> settings,
+        CardanoService cardanoService)
     {
         _logger = logger;
         _dbContextFactory = dbContextFactory;
+        _settings = settings.Value;
+        _cardanoService = cardanoService;
     }
     public async Task ReduceAsync(OuraStakeDelegationEvent stakeDelegationEvent)
     {
@@ -23,20 +35,25 @@ public class DelegatorByEpochReducer : OuraReducerBase
             stakeDelegationEvent.StakeDelegation is not null &&
             stakeDelegationEvent.StakeDelegation.Credential is not null &&
             stakeDelegationEvent.StakeDelegation.PoolHash is not null &&
-            stakeDelegationEvent.Context.Slot is not null)
+            stakeDelegationEvent.Context.TxIdx is not null &&
+            stakeDelegationEvent.Context.TxHash is not null)
         {
             Block? block = await _dbContext.Block.Where(b => b.BlockHash == stakeDelegationEvent.Context.BlockHash).FirstOrDefaultAsync();
 
             if (block is null) return;
 
-            string stakeAddress = string.IsNullOrEmpty(stakeDelegationEvent.StakeDelegation.Credential.AddrKeyHash) ?
+            string stakeKeyHash = string.IsNullOrEmpty(stakeDelegationEvent.StakeDelegation.Credential.AddrKeyHash) ?
                 stakeDelegationEvent.StakeDelegation.Credential.Scripthash :
                 stakeDelegationEvent.StakeDelegation.Credential.AddrKeyHash;
 
+            string stakeAddress = AddressUtility.GetRewardAddress(stakeKeyHash.HexToByteArray(), _settings.NetworkType).ToString();
+            string poolId = _cardanoService.PoolHashToBech32(stakeDelegationEvent.StakeDelegation.PoolHash);
+
             DelegatorByEpoch? entry = await _dbContext.DelegatorByEpoch.Include(dbe => dbe.Block)
                 .Where((dbe) => dbe.StakeAddress == stakeAddress &&
-                        dbe.Slot == block.Slot &&
-                        dbe.PoolHash == stakeDelegationEvent.StakeDelegation.PoolHash)
+                    dbe.TxHash == stakeDelegationEvent.Context.TxHash &&
+                    dbe.PoolId == poolId &&
+                    dbe.TxIndex == (ulong)stakeDelegationEvent.Context.TxIdx)
                 .FirstOrDefaultAsync();
 
             if (entry is not null &&
@@ -46,11 +63,10 @@ public class DelegatorByEpochReducer : OuraReducerBase
 
             await _dbContext.DelegatorByEpoch.AddAsync(new()
             {
-                StakeAddress = string.IsNullOrEmpty(stakeDelegationEvent.StakeDelegation.Credential.AddrKeyHash) ?
-                    stakeDelegationEvent.StakeDelegation.Credential.Scripthash :
-                    stakeDelegationEvent.StakeDelegation.Credential.AddrKeyHash,
-                PoolHash = stakeDelegationEvent.StakeDelegation.PoolHash,
-                Slot = block.Slot,
+                StakeAddress = stakeAddress,
+                PoolId = poolId,
+                TxHash = stakeDelegationEvent.Context.TxHash,
+                TxIndex = (ulong)stakeDelegationEvent.Context.TxIdx,
                 Block = block
             });
 
@@ -60,10 +76,10 @@ public class DelegatorByEpochReducer : OuraReducerBase
 
     public async Task RollbackAsync(Block rollbackBlock)
     {
-        _logger.LogInformation("DelegatorByEpoch Rollback...");
         using ConclaveSinkDbContext _dbContext = await _dbContextFactory.CreateDbContextAsync();
         IEnumerable<DelegatorByEpoch>? delegatorByEpoch = _dbContext.DelegatorByEpoch.Include(dbe => dbe.Block)
             .Where(dbe => dbe.Block!.BlockHash == rollbackBlock.BlockHash);
+        if (delegatorByEpoch is null) return;
         _dbContext.DelegatorByEpoch.RemoveRange(delegatorByEpoch);
         await _dbContext.SaveChangesAsync();
     }
